requirements.txt
asgiref==3.11.1
debugpy==1.8.20
Django==5.2.13
django-debug-toolbar==6.3.0
djangorestframework==3.17.1
pillow==12.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
djangorestframework-simplejwt
